Remove Cypress e2e tests
The selected mapping library uses WebGL to render map elements. This
means it is not possible for a DOM testing library to select and assert
on map markers, shapes etc.

Cypress, and other browser testing tools, are therefore limited in
their ability to provide confidence.

I've decided to remove Cypress as the complexity overhead now exceeds
any benefits.